refactor: adapt to v9.0.0
