--- core.d.ts
@@ -1,16 +1,0 @@
-export * from './minimal';
-export {
-  Canonizer,
-  CanonizerUtil,
-  ConformerGenerator,
-  DruglikenessPredictor,
-  DrugScoreCalculator,
-  ForceFieldMMFF94,
-  IInitializeConformersOptions,
-  IParameterizedString,
-  MoleculeProperties,
-  ReactionEncoder,
-  Reactor,
-  Transformer,
-  ToxicityPredictor,
-} from './types';
