+++ lib/index.js
@@ -13,6 +13,7 @@ export default OCL;
 export const CanvasEditor = OCL.CanvasEditor;
 export const registerCustomElement = OCL.registerCustomElement;
 
+export const Canonizer = OCL.Canonizer;
 export const CanonizerUtil = OCL.CanonizerUtil;
 export const ConformerGenerator = OCL.ConformerGenerator;
 export const DrugScoreCalculator = OCL.DrugScoreCalculator;
